Unificado, academicTerm funcionando con acls.
Creado algún metodo para ver los los grupos de un usuario (ya sea student o professor)

Cambiado el logo-header

+++ src/main/java/com/example/tfg/service/implementation/CourseServiceImp.java
@@ -1,5 +1,6 @@
 package com.example.tfg.service.implementation;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
@@ -9,12 +10,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.example.tfg.classes.ResultClass;
 import com.example.tfg.domain.AcademicTerm;
 import com.example.tfg.domain.Course;
 import com.example.tfg.repository.CourseDao;
 import com.example.tfg.service.AcademicTermService;
 import com.example.tfg.service.ActivityService;
 import com.example.tfg.service.CourseService;
+import com.example.tfg.service.GroupService;
 
 @Service
 public class CourseServiceImp implements CourseService {
@@ -24,6 +27,9 @@ public class CourseServiceImp implements CourseService {
 
 	@Autowired
 	ActivityService serviceActivity;
+	
+	@Autowired
+	GroupService serviceGroup;
 
 	// @Autowired
 	// private DegreeService serviceDegree;
@@ -33,26 +39,33 @@ public class CourseServiceImp implements CourseService {
 
 	@PreAuthorize("hasRole('ROLE_ADMIN')")	
 	@Transactional(readOnly = false)
-	public boolean addCourse(Course course, Long id_academic) {
+	public ResultClass<Boolean> addCourse(Course course, Long id_academic) {
 		
-		AcademicTerm academic = serviceAcademicTerm
-				.getAcademicTerm(id_academic);
-		course.setAcademicTerm(academic);
-		Course existCourse = daoCourse.exist(course);
-		if(existCourse == null){
-//			course.setAcademicTerm(academic);
-			academic.getCourses().add(course);
-			return daoCourse.addCourse(course);				
-		}
-		else if(existCourse.isDeleted()){
-			existCourse.setAcademicTerm(academic);
-			academic.getCourses().add(existCourse);
-			existCourse.setDeleted(false);
-			existCourse.setSubject(course.getSubject());
-			return daoCourse.saveCourse(existCourse);
-		}
-		else return false;
+		course.setAcademicTerm(serviceAcademicTerm.getAcademicTerm(id_academic));
+		Course courseExists = daoCourse.exist(course);
+		ResultClass<Boolean> result = new ResultClass<Boolean>();
 
+		if( courseExists != null){
+			result.setHasErrors(true);
+			Collection<String> errors = new ArrayList<String>();
+			errors.add("Code already exists");
+
+			if (courseExists.isDeleted()){
+				result.setElementDeleted(true);
+				errors.add("Element is deleted");
+
+			}
+			result.setE(false);
+			result.setErrorsList(errors);
+		}
+		else{
+		
+			boolean r = daoCourse.addCourse(course);
+			if (r) 
+				result.setE(true);
+		}
+		return result;		
+	
 	}
 
 	@PreAuthorize("hasRole('ROLE_USER')")
@@ -61,23 +74,43 @@ public class CourseServiceImp implements CourseService {
 		return daoCourse.getAll();
 	}
 
+
 	//yo lo borraria no tiene sentido en este caso el modify
 	@PreAuthorize("hasRole('ROLE_ADMIN')")	
 	@Transactional(readOnly = false)
-	public boolean modifyCourse(Course course, Long id_academic, Long id_course) {
-		Course courseModify = daoCourse.getCourse(id_course);
+	public ResultClass<Boolean> modifyCourse(Course course, Long id_academic, Long id_course) {
+		ResultClass<Boolean> result = new ResultClass<Boolean>();
 		
 		course.setAcademicTerm(serviceAcademicTerm.getAcademicTerm(id_academic));
-		Course existModify = daoCourse.exist(course);
 		
+		Course modifyCourse = daoCourse.getCourse(id_course);
 		
-		if (existModify == null){
-				courseModify.setSubject(course.getSubject());
-				return daoCourse.saveCourse(courseModify);
+		Course courseExists = daoCourse.exist(course);
+		
+		if(!course.getSubject().equals(modifyCourse.getSubject()) && 
+				courseExists != null){
+			result.setHasErrors(true);
+			Collection<String> errors = new ArrayList<String>();
+			errors.add("New code already exists");
+
+			if (courseExists.isDeleted()){
+				result.setElementDeleted(true);
+				errors.add("Element is deleted");
+
+			}
+			result.setErrorsList(errors);
+			result.setE(false);
 		}
+		else{
+			modifyCourse.setSubject(course.getSubject());
+			boolean r = daoCourse.saveCourse(modifyCourse);
+			if (r) 
+				result.setE(true);
+		}
+		return result;
 
 			
-		else return false;
+
 	}
 
 	@PreAuthorize("hasRole('ROLE_USER')")
@@ -105,9 +138,9 @@ public class CourseServiceImp implements CourseService {
 
 	@PreAuthorize("hasRole('ROLE_ADMIN')")	
 	public boolean deleteCoursesFromAcademic(AcademicTerm academic) {
-		boolean deleted = serviceActivity.deleteActivitiesFromCourses(academic.getCourses());
-		
-		if (deleted)
+		boolean deleteActivities = serviceActivity.deleteActivitiesFromCourses(academic.getCourses());
+		boolean deleteGroups = serviceGroup.deleteGroupsFromCourses(academic.getCourses());
+		if (deleteActivities && deleteGroups)
 			return daoCourse.deleteCoursesFromAcademic(academic);
 		else return false;
 
@@ -115,10 +148,14 @@ public class CourseServiceImp implements CourseService {
 	
 
 
+	@Transactional(readOnly = true)
 	@PreAuthorize("hasRole('ROLE_USER')")	
 	public Course getCourseAll(Long id) {
 		Course c = daoCourse.getCourse(id);
 		c.setActivities(serviceActivity.getActivitiesForCourse(id));
+		c.setGroups(serviceGroup.getGroupsForCourse(id));
+		
+	
 		return c;
 	}
 
@@ -133,9 +170,44 @@ public class CourseServiceImp implements CourseService {
 	public boolean deleteCourses(Collection<AcademicTerm> academicList) {
 		Collection<Course> coursesList = daoCourse.getCoursesFromListAcademic(academicList);
 		boolean deleteActivities = serviceActivity.deleteActivitiesFromCourses(coursesList);
-		if (deleteActivities)
+		boolean deleteGroups = serviceGroup.deleteGroupsFromCourses(coursesList);
+		if (deleteActivities && deleteGroups)
 		return daoCourse.deleteCourses(academicList);
 		else return false;
 	}
 
+
+	public boolean modifyCourse(Course course) {
+
+		return daoCourse.saveCourse(course);
+	}
+
+	@PreAuthorize("hasRole('ROLE_ADMIN')")	
+	@Transactional(readOnly = false)
+	public ResultClass<Boolean> unDeleteCourse(Course course) {
+		Course c = daoCourse.exist(course);
+		ResultClass<Boolean> result = new ResultClass<Boolean>();
+		if(c == null){
+			result.setHasErrors(true);
+			Collection<String> errors = new ArrayList<String>();
+			errors.add("Code doesn't exist");
+			result.setErrorsList(errors);
+
+		}
+		else{
+			if(!c.isDeleted()){
+				Collection<String> errors = new ArrayList<String>();
+				errors.add("Code is not deleted");
+				result.setErrorsList(errors);
+			}
+
+			c.setDeleted(false);
+			c.setSubject(course.getSubject());
+			boolean r = daoCourse.saveCourse(c);
+			if(r) 
+				result.setE(true);	
+
+		}
+		return result;
+	}
 }
